质量检验合格数量计算（机加工、零件质检）

+++ src/main/java/com/deepsoft/haolifa/dao/repository/InspectMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface InspectMapper {
-    long countByExample(InspectExample example);
+    int countByExample(InspectExample example);
 
     int deleteByExample(InspectExample example);
